Got edit, delete, and clear buttons working properly
