+++ src/javascripts/components/addComments.js
@@ -2,12 +2,11 @@ import util from '../helpers/util';
 
 import seedData from './seedData';
 
-
 const commentAvatar = 'https://via.placeholder.com/150';
 let commentNum = 1;
 let commentCollection = [];
-// let seedArrayEmpty = [];
 
+// This builds the comment cards and prints them to the DOM
 const messageBuilder = (commentArray) => {
   let domString = '';
   commentArray.forEach((comment) => {
@@ -19,19 +18,21 @@ const messageBuilder = (commentArray) => {
     domString += `<input id="${comment.id}" class="newCommentEntry" value="${comment.comment}">`;
     domString += `<button id="${comment.id}" class="btn btn-primary updateCommentBtn">Update</button>`;
     domString += '</div>';
-    domString += `<button id="${comment.id}" class="btn btn-danger deleteButton">Delete</button>`;
+    domString += `<button id="${comment.id}" class="btn btn-danger deleteBtn">Delete</button>`;
     domString += `<button id="${comment.id}" class="btn btn-primary editBtn">Edit</button>`;
     domString += '</div>';
   });
   util.printToDom('container', domString);
 };
 
+// This only shows the most recent 20 comments
 const limitMessages = () => {
   if (commentCollection.length > 20) {
     commentCollection.shift();
   }
 };
 
+// This adds a comment to the DOM by taking user input and adding new object to commentCollection
 const addComment = () => {
   const inputName = document.getElementById('userName');
   const inputComment = document.getElementById('userComment');
@@ -51,7 +52,105 @@ const addComment = () => {
   inputComment.value = '';
 };
 
+// This takes the updated comment text and plugs it into the innerHTML of the comment to edit
+// It then hides the input and update button fields
+const updateCommentText = (targetId) => {
+  const allUpdateButtons = Array.from(document.getElementsByClassName('updateCommentBtn'));
+  const allInputs = Array.from(document.getElementsByClassName('newCommentEntry'));
+  let replacementText = '';
+  allInputs.forEach((input) => {
+    if (input.id === targetId) {
+      replacementText = input.value;
+    }
+  });
+  commentCollection.forEach((message) => {
+    if (message.id === targetId) {
+      message.comment = replacementText;
+    }
+    return commentCollection;
+  });
+  allInputs.forEach((input) => {
+    if (input.id === targetId) {
+      console.error(input.id, targetId);
+      input.style.display = 'none';
+    }
+  });
+  allUpdateButtons.forEach((button) => {
+    if (button.id === targetId) {
+      button.style.display = 'none';
+    }
+  });
+  messageBuilder(commentCollection);
+};
 
+// This adds a listener to the update button on the comment card that will run updateCommentText
+const addUpdateBtnListener = (myId) => {
+  const allUpdateButtons = Array.from(document.getElementsByClassName('updateCommentBtn'));
+  allUpdateButtons.forEach((button) => {
+    button.addEventListener('click', () => {
+      updateCommentText(myId);
+    });
+  });
+};
+
+// This runs when any edit button is clicked. Causes an input field and update button to appear.
+// It also adds a listener to the update button
+const editText = (e) => {
+  const allCommentTexts = Array.from(document.getElementsByClassName('commentText'));
+  const allUpdateButtons = Array.from(document.getElementsByClassName('updateCommentBtn'));
+  const allInputs = Array.from(document.getElementsByClassName('newCommentEntry'));
+  const buttonId = e.target.id;
+  allCommentTexts.forEach((comment) => {
+    if (comment.id === buttonId) {
+      console.error(comment.id);
+      allInputs.forEach((input) => {
+        if (input.id === buttonId) {
+          input.style.display = 'inline';
+        }
+      });
+      allUpdateButtons.forEach((button) => {
+        if (button.id === buttonId) {
+          button.style.display = 'inline';
+        }
+      });
+    }
+  });
+  addUpdateBtnListener(buttonId);
+};
+
+// This removes a comment from commentCollection and reprints commentCollection to the DOM
+const deleteComment = (e) => {
+  commentCollection.forEach((comment) => {
+    if (comment.id === e.target.id) {
+      const index = commentCollection.indexOf(comment);
+      commentCollection.splice(index, 1);
+      messageBuilder(commentCollection);
+    }
+  });
+};
+
+// This clears all messages from the DOM
+const clearComments = () => {
+  commentCollection = [];
+  messageBuilder(commentCollection);
+};
+
+// This adds listeners to the 'edit' and 'delete' buttons on cards as well as the 'clear' button
+const addEditBtnListeners = () => {
+  document.addEventListener('click', (e) => {
+    if (e.target.classList.contains('editBtn')) {
+      editText(e);
+    }
+    if (e.target.id === 'clearBtn') {
+      clearComments();
+    }
+    if (e.target.classList.contains('deleteBtn')) {
+      deleteComment(e);
+    }
+  });
+};
+
+// This listens for the user to press enter or click after entering a comment and prints it
 const addCommentEvents = () => {
   const addCommentBtn = document.getElementById('addCommentBtn');
   const userComment = document.getElementById('userComment');
@@ -69,6 +168,7 @@ const addCommentEvents = () => {
   });
 };
 
+// This takes the seed data and prints it as comments
 const getData = () => {
   seedData.getSeedData()
     .then((response) => {
@@ -81,6 +181,4 @@ const getData = () => {
     });
 };
 
-const returnCommentCollection = () => commentCollection;
-
-export default { addCommentEvents, getData, returnCommentCollection };
+export default { addCommentEvents, getData, addEditBtnListeners };
