+++ src/javascripts/main.js
@@ -1,7 +1,6 @@
 import themesModal from './components/themes-modal/themes-modal';
 
 import addCommentEvents from './components/addComments';
-import editButtonListeners from './components/editComment/editButtonListeners';
 
 import 'bootstrap';
 import '../styles/main.scss';
@@ -10,7 +9,7 @@ const init = () => {
   themesModal.modalControlsListeners();
   addCommentEvents.getData();
   addCommentEvents.addCommentEvents();
-  editButtonListeners.addEditBtnListeners();
+  addCommentEvents.addEditBtnListeners();
 };
 
 init();
