+++ src/styles/_colors.scss
@@ -1,8 +1,16 @@
+$fire-navbar-background: url("../../images/fire.png");
+$fire-page-background: rgb(54, 12, 12);
 $fire-background: #D2431D;
 $fire-font: #F2CE6D;
+$ice-navbar-background: url("../../images/ice.jpg");
+$ice-page-background: rgb(2, 2, 29);
 $ice-background: #8EE5EB;
 $ice-font: black;
+$stone-navbar-background: url("../../images/stone.jpg");
+$stone-page-background: rgb(29, 28, 28);
 $stone-background: #343A3A;
 $stone-font: white;
+$original-navbar-background: url("../../images/default.png");
+$original-page-background: rgb(3, 44, 44);
 $original-background: teal;
 $original-font: black;
