+++ src/styles/main.scss
@@ -40,7 +40,7 @@ body {
   flex-direction: column;
   justify-content: space-around;
   height: 300px;
-  background-image: url("../../images/theWall.jpg");
+  background-image: url("../../images/default.png");
   background-position: center;
   background-size: cover;
   position: static;
